added logboek component
